Fix scaling in the IDE once again
Turns out the previous change was wrong — the issue was mostly due to us
taking some JBDimensions and not un-scaling them before using their
values.

To better match the line width in Swing, I also tweaked the border
modifier logic to not round up line widths, as that was causing lines
to be drawn thicker than they should. This is not a full fix, since
Swing does some tweaks in DarculaNewUIUtils#paintRectangle, and it
probably just draws things slightly differently.

However, this does considerably improve things at most IDE zoom levels,
at the cost of some minor fuzziness at 110% zoom.
