+++ ide-laf-bridge/src/main/kotlin/org/jetbrains/jewel/bridge/theme/SwingBridgeTheme.kt
@@ -8,7 +8,7 @@ import androidx.compose.ui.platform.LocalDensity
 import com.intellij.openapi.components.service
 import org.jetbrains.jewel.bridge.BridgePainterHintsProvider
 import org.jetbrains.jewel.bridge.SwingBridgeService
-import org.jetbrains.jewel.bridge.retrieveIdeaDensity
+import org.jetbrains.jewel.bridge.scaleDensityWithIdeScale
 import org.jetbrains.jewel.foundation.ExperimentalJewelApi
 import org.jetbrains.jewel.ui.ComponentStyling
 import org.jetbrains.jewel.ui.painter.LocalPainterHintsProvider
@@ -29,7 +29,7 @@ public fun SwingBridgeTheme(content: @Composable () -> Unit) {
     ) {
         CompositionLocalProvider(
             LocalPainterHintsProvider provides BridgePainterHintsProvider(themeData.themeDefinition.isDark),
-            LocalDensity provides retrieveIdeaDensity(LocalDensity.current),
+            LocalDensity provides scaleDensityWithIdeScale(LocalDensity.current),
         ) {
             content()
         }
